Rename `task-list-generator` unit test files to match v4/v5 change
